notes view make sense

+++ main/views.py
@@ -17,8 +17,8 @@ from rest_framework import status
 from main import models, tables, filters, forms
 from dateutil.relativedelta import relativedelta
 
-#TODO zrobic multiple ip
 #TODO cos przejscie z activity do transactions multiple tags ucinalo
+#TODO Tags multiple header value 
 #TODO wrzucic na azure
 #TODO Godziny w activity 10:13 - 11:13
 #TODO Przyciski w activity
@@ -27,11 +27,11 @@ from dateutil.relativedelta import relativedelta
 #TODO dodac paginacje dla visitors spytac sie o ilosc danych czy warrto robic pginacje dla tagow czy tez serwera
 
 
-def all_notes_form_view(request,id=None,ip=None,transaction=None):
+def all_notes_form_view(request,id=None,visitor_ip=None,transaction=None):
     if request.method == "GET":
         if not id:
             initial = {
-                "ip" : ip,
+                "visitor_ip" : visitor_ip,
                 "transaction" : transaction
             }
             form = forms.NoteForm(initial=initial)
@@ -46,7 +46,7 @@ def all_notes_form_view(request,id=None,ip=None,transaction=None):
     if request.method == "POST":
         if not id:
             initial = {
-                "ip" : ip,
+                "visitor_ip" : visitor_ip,
                 "transaction" : transaction
             }
             note = models.LogsNote(**initial)
@@ -61,31 +61,32 @@ def all_notes_form_view(request,id=None,ip=None,transaction=None):
             if transaction:
                 t = [models.Transaction.objects.get(pk=transaction)]
             else:   
-                t = models.Transaction.objects.filter(Q(name__header_name=settings.VISITOR_IP) & Q(value__header_value=ip))
+                t = models.Transaction.objects.filter(Q(name__header_name=settings.VISITOR_IP) & Q(value__header_value=visitor_ip))
             logsnote.transactions.add(*t)
             return HttpResponseRedirect(reverse("all_notes"))
         return render(request, "all_notes_form.html", { "form" : form })
 
 def all_notes_view(request):
     queryset = models.LogsNote.objects.all()
-    initial = {"select":"search"}
-    if request.method == "POST":
-        if request.POST.get("select") == "delete_selected" \
-        and request.POST.__contains__("selected_notes"):
+    initial = {"select":"search"}    
+    print(request.GET)
+    if request.method == "POST"and request.POST.__contains__("selected_notes"):
             notes_to_delete = request.POST.getlist("selected_notes")
             models.LogsNote.objects.filter(id__in=notes_to_delete).delete()
-        elif request.POST.get("select") == "search" and request.POST.get("title"):
-            queryset = models.LogsNote.objects.filter(title__istartswith=request.POST.get("title"))
-            initial["title"] = request.POST.get("title")
-    
-    form = forms.NotesActionSelectForm(initial=initial)
-    table = tables.NotesTable(queryset, order_by="-id") 
+        # elif request.POST.get("select") == "search" and request.POST.get("title"):
+        #     queryset = models.LogsNote.objects.filter(title__istartswith=request.POST.get("title"))
+        #     initial["title"] = request.POST.get("title")
+        #     filter = filters.NoteFilter(request.POST,queryset=queryset)
+        #     queryset
+    # form = forms.NotesActionSelectForm(initial=initial)
+    filter = filters.NoteFilter(request.GET,queryset=queryset)
+    table = tables.NotesTable(filter.qs, order_by="-id") 
     table.paginate(page=request.GET.get("page", 1), per_page=5)
     return render(request, "all_notes.html",  {
-        "form" : form,
+        "form" : filter.form,
         "table": table
     })
-
+    return render(request,"all_notes.html")
 def activity_view(request):
     form = filters.ChartFilter(request.GET).form
     return render(request, "activity.html",  {
@@ -93,14 +94,14 @@ def activity_view(request):
         "tag_field" : "assigned_tags",
     })
 
-def transactions_detail_view(request, transaction=1,ip=1):
+def transactions_detail_view(request, transaction=1,visitor_ip=1):
     t = models.Transaction.objects.get(pk=transaction)
-    ip = t.logslog_set.filter(value=ip).first().value.header_value
+    visitor_ip = t.logslog_set.filter(value=visitor_ip).first().value.header_value
     table = tables.TransactionsDetailTable(t.logslog_set.all())    
     return render(request, "transactions_detail.html", {
         "table":table,
         "transaction": transaction,
-        "ip": ip
+        "visitor_ip": visitor_ip
     })
 
 def tags_form_view(request, id=None):
